+++ tests/test_FLA/test_chunk_gated_delta_rule.py
@@ -38,9 +38,7 @@ def _install_triton_allocator():
         return
 
     def _alloc(size: int, _alignment: int, _stream: int | None):
-        return torch.empty(
-            (size,), dtype=torch.uint8, device=flaggems_vllm.device
-        )
+        return torch.empty((size,), dtype=torch.uint8, device=flaggems_vllm.device)
 
     triton.set_allocator(_alloc)
     _TRITON_ALLOCATOR_READY = True
@@ -96,22 +94,18 @@ def _make_inputs(
     dtype: torch.dtype,
     head_first: bool,
     non_contiguous: bool = False,
-) -> tuple[
-    torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor
-]:
+) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]:
     device = flaggems_vllm.device
-    q_data = torch.randn(B, T, Hg, K, device=device, dtype=torch.float32).to(
-        dtype
-    )
+    q_data = torch.randn(B, T, Hg, K, device=device, dtype=torch.float32).to(dtype)
     k_data = F.normalize(
         torch.randn(B, T, Hg, K, device=device, dtype=torch.float32),
         p=2.0,
         dim=-1,
         eps=1e-6,
     ).to(dtype)
-    v_data = (
-        0.125 * torch.randn(B, T, H, V, device=device, dtype=torch.float32)
-    ).to(dtype)
+    v_data = (0.125 * torch.randn(B, T, H, V, device=device, dtype=torch.float32)).to(
+        dtype
+    )
     if non_contiguous:
         q = _strided_last_dim(q_data)
         k = _strided_last_dim(k_data)
@@ -179,9 +173,7 @@ def _reference_chunk_gated_delta_rule(
         spans = [(b, b, 0, T) for b in range(B)]
     else:
         cu_cpu = cu_seqlens.detach().cpu().tolist()
-        spans = [
-            (0, n, cu_cpu[n], cu_cpu[n + 1]) for n in range(len(cu_cpu) - 1)
-        ]
+        spans = [(0, n, cu_cpu[n], cu_cpu[n + 1]) for n in range(len(cu_cpu) - 1)]
 
     for batch_idx, state_idx, start, end in spans:
         if initial_state is None:
@@ -228,9 +220,7 @@ def _assert_close(
     )
 
 
-@pytest.mark.parametrize(
-    "dtype", [torch.float32, torch.float16, torch.bfloat16]
-)
+@pytest.mark.parametrize("dtype", [torch.float32, torch.float16, torch.bfloat16])
 @pytest.mark.parametrize("head_first", [True, False])
 def test_chunk_gated_delta_rule_matches_reference_without_final_state(
     dtype, head_first
@@ -560,9 +550,7 @@ def test_chunk_gated_delta_rule_direct_path_shape_variants(case):
 
     _assert_close(actual, expected, case["dtype"])
     if output_final_state:
-        _assert_close(
-            actual_final, expected_final, case["dtype"], final_state=True
-        )
+        _assert_close(actual_final, expected_final, case["dtype"], final_state=True)
     else:
         assert actual_final is None
         assert expected_final is None
@@ -629,9 +617,7 @@ def test_chunk_gated_delta_rule_zero_values_extreme_gates_return_zero_state():
         head_first=False,
     )
 
-    torch.testing.assert_close(
-        actual, torch.zeros_like(actual), atol=0, rtol=0
-    )
+    torch.testing.assert_close(actual, torch.zeros_like(actual), atol=0, rtol=0)
     torch.testing.assert_close(
         actual_final,
         torch.zeros_like(actual_final),
